Restructure run-time folder
