+++ runtime/src/common.h
